Convert to streams
Without streams, data is processed as quickly as possible and queued up as
async events waiting for processing (which are processed at the rate of the
receiving pipe). If the receiving pipe is to SSD, this is fine. If it is to
gzip, then we run out of heap in node.js very quickly. Streams apply
backpressure to the input and slow down reading to only process at the rate
the output pipe is able to accept.